docs: add sentinel database table and column docs
This is the markdown generated by https://observablehq.com/d/7541c525585b1d5d. Adding it here also for visibility and speedy access.

Not everyone has access to the team observable and it takes time to generate due to the many queries that have to be sent to get the comment information.

The markdown for this file is generated and logged out at the bottom of the observable so updating this should be as easy as copying it over from the github action. Maybe someone wants to automate that with a github action in the future.